+++ frontend/src/routes/[id]/+page.svelte
@@ -1,0 +1,7 @@
+<script>
+	import { page } from '$app/stores';
+</script>
+
+Your id token is:
+
+{$page.params.id}
